修bug
修复氪金计算条件判断语句和赋值语句不相同导致的index out of range异常

+++ GIGachaSupport.py
@@ -245,7 +245,7 @@ def sim(IntertwinedFateNum, Discounts, KrAu, KrAuEx, ExpectedCharacterNum, Chara
     # 首充计算
     FirstEx = int(KrAu / 8)
     if FirstEx:
-        if len(percentlist) > IntertwinedFateNum + FirstEx / Discounts:
+        if len(percentlist) > int((IntertwinedFateNum + FirstEx) / Discounts):
             position = int((IntertwinedFateNum + FirstEx) / Discounts)
             possibility = percentlist[position]
             text = f'首充:{position}抽,成功率{possibility * 100:.1f}%'
@@ -259,7 +259,7 @@ def sim(IntertwinedFateNum, Discounts, KrAu, KrAuEx, ExpectedCharacterNum, Chara
     if KrAuEx >= 648:
         cols = generate_gradient_colors(int(KrAuEx / 648), 'gold', 'red')  # 生成渐变色标识
         for i in range(int(KrAuEx / 648)):
-            if len(percentlist) > (IntertwinedFateNum + 50.5 * (i + 1) + FirstEx) / Discounts:
+            if len(percentlist) > int((IntertwinedFateNum + 50.5 * (i + 1) + FirstEx) / Discounts):
                 position = int((IntertwinedFateNum + 50.5 * (i + 1) + FirstEx) / Discounts)
                 possibility = percentlist[position]
                 text = f'氪{i + 1}单:{position}抽,成功率{possibility * 100:.1f}%'
